requirements.txt
Flask==3.1.0
bcrypt==4.2.0
pycryptodome==3.21.0